pulled out useCommunitiesFeatured.js
