+++ src/components/FeaturedCommunities.js
@@ -1,40 +1,13 @@
+import React from 'react';
 import { CircularProgress } from '@material-ui/core';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
-import React, { useEffect, useState } from 'react';
-import endpoints from '../constants/endpoints';
-import routes from '../constants/routes';
-import { trackException } from '../services/telemetry.service';
 import ErrorSnackbar from './ErrorSnackbar';
 import FeaturedPost from './FeaturedPost';
+import useCommunitiesFeatured from '../hooks/useCommunitiesFeatured';
 
 export default function FeaturedCommunities() {
-  const [error, setError] = useState(null);
-  const [isLoaded, setLoaded] = useState(false);
-  const [communities, setCommunities] = useState([]);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      await fetch(endpoints.communitiesFeatured)
-        .then(res => res.json())
-        .then(
-          result => {
-            const s = result.map(x => ({
-              ...x,
-              path: `${routes.communities.path}/${x.slug}`,
-            }));
-            setCommunities(s);
-            setLoaded(true);
-          },
-          e => {
-            setError(e);
-            setLoaded(true);
-            trackException(e);
-          },
-        );
-    };
-    fetchData();
-  }, []);
+  const { error, isLoaded, communities } = useCommunitiesFeatured();
 
   return (
     <>
