+++ src/hooks/useCommunitiesFeatured.js
@@ -1,0 +1,39 @@
+import { useEffect, useState } from 'react';
+import endpoints from '../constants/endpoints';
+import routes from '../constants/routes';
+import { trackException } from '../services/telemetry.service';
+
+export default function useCommunitiesFeatured() {
+  const [error, setError] = useState(null);
+  const [isLoaded, setLoaded] = useState(false);
+  const [communities, setCommunities] = useState([]);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      await fetch(endpoints.communitiesFeatured)
+        .then(res => res.json())
+        .then(
+          result => {
+            const s = result.map(x => ({
+              ...x,
+              path: `${routes.communities.path}/${x.slug}`,
+            }));
+            setCommunities(s);
+            setLoaded(true);
+          },
+          e => {
+            setError(e);
+            setLoaded(true);
+            trackException(e);
+          },
+        );
+    };
+    fetchData();
+  }, []);
+
+  return {
+    error,
+    isLoaded,
+    communities,
+  };
+}
